Refactor setup to modular plugins
