+++ erki-kickstart/lua/custom/plugins/markdown-preview.lua
@@ -1,0 +1,6 @@
+return {
+  "iamcco/markdown-preview.nvim",
+  cmd = { "MarkdownPreviewToggle", "MarkdownPreview", "MarkdownPreviewStop" },
+  ft = { "markdown" },
+  build = function() vim.fn["mkdp#util#install"]() end,
+}
