+++ erki-kickstart/lua/custom/plugins/noice.lua
@@ -1,0 +1,9 @@
+return {
+  "folke/noice.nvim",
+  event = "VeryLazy",
+  opts = {},
+  dependencies = {
+    "MunifTanjim/nui.nvim",
+    "rcarriga/nvim-notify",
+  },
+}
