+++ erki-kickstart/lua/custom/plugins/snacks.lua
@@ -1,0 +1,15 @@
+return {
+  "folke/snacks.nvim",
+  priority = 1000,
+  lazy = false,
+  opts = {
+    lazygit = {},
+    notifier = {
+      enabled = true,
+      timeout = 2000,
+    },
+  },
+  keys = {
+    { "<leader>gg", function() Snacks.lazygit() end, desc = "Lazygit" },
+  },
+}
