+++ erki-kickstart/lua/custom/plugins/yazi.lua
@@ -1,0 +1,11 @@
+return {
+  "DreamMaoMao/yazi.nvim",
+  dependencies = {
+    "nvim-telescope/telescope.nvim",
+    "nvim-lua/plenary.nvim",
+  },
+
+  keys = {
+    { "<leader>gy", "<cmd>Yazi<CR>", desc = "Toggle Yazi" },
+  },
+}
